Changed getElapsedTime() to return an int since it's guaranteed the be less than duration, which is an int
git-svn-id: https://svn.apache.org/repos/asf/incubator/pivot/trunk@753994 13f79535-47bb-0310-9956-ffa450edef68

+++ wtk/src/pivot/wtk/effects/Transition.java
@@ -191,8 +191,8 @@ public abstract class Transition {
      * Returns the amount of time that has passed since the transition
      * was started.
      */
-    public long getElapsedTime() {
-        return currentTime - startTime;
+    public int getElapsedTime() {
+        return (int)(currentTime - startTime);
     }
 
     /**
